+++ src/bunit.core/Rendering/BunitComponentActivator.cs
@@ -1,17 +1,16 @@
 #if NET5_0_OR_GREATER
-using System;
-using System.Diagnostics.CodeAnalysis;
-using Microsoft.AspNetCore.Components;
 
 namespace Bunit.Rendering;
 
 internal class BunitComponentActivator : IComponentActivator
 {
 	private readonly ComponentFactoryCollection factories;
+	private readonly IComponentActivator componentActivator;
 
-	public BunitComponentActivator(ComponentFactoryCollection factories)
+	public BunitComponentActivator(ComponentFactoryCollection factories, IComponentActivator? externalComponentActivator)
 	{
 		this.factories = factories ?? throw new ArgumentNullException(nameof(factories));
+		this.componentActivator = externalComponentActivator ?? DefaultComponentActivator.Instance;
 	}
 
 	public IComponent CreateInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type componentType)
@@ -39,7 +38,18 @@ internal class BunitComponentActivator : IComponentActivator
 			}
 		}
 
-		return (IComponent)Activator.CreateInstance(componentType)!;
+		return componentActivator.CreateInstance(componentType);
+	}
+
+	private sealed class DefaultComponentActivator : IComponentActivator
+	{
+		public static IComponentActivator Instance { get; } = new DefaultComponentActivator();
+
+		/// <inheritdoc />
+		public IComponent CreateInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] Type componentType)
+		{			
+			return (IComponent)Activator.CreateInstance(componentType)!;
+		}
 	}
 }
 #endif